Cholak/rocket

+++ spbstu_ss/src/main.rs
@@ -13,32 +13,37 @@ use spbstu_ss::controllers::memberships_controller::{MembershipsController, Memb
 use std::convert::Infallible;
 
 use auth::*;
-use spbstu_ss::models::group_model::{Group, NewGroup, UpdatedGroup};
 use spbstu_ss::controllers::users_controller::*;
 use spbstu_ss::controllers::groups_controller::*;
-use spbstu_ss::models::membership_model::{Membership, NewMembership, UpdatedMembership};
 use spbstu_ss::models::user_model::{NewUser, UpdatedUser, User};
-use spbstu_ss::models::{Claims, LoginRequest, LoginResponse};
+use spbstu_ss::models::group_model::{Group, NewGroup, UpdatedGroup};
+use spbstu_ss::models::membership_model::{Membership, NewMembership, UpdatedMembership};
+use spbstu_ss::models::auth_model::{Claims, LoginRequest, LoginResponse};
 
 #[get("/hello/<name>/<age>")]
 fn hello(name: String, age: u8) -> String {
     format!("Hello? {} year old named {}!", age, name)
 }
 
-#[post("/login", format = "json", data = "<name>")]
-pub fn login(name: Json<LoginRequest>) -> Result<Json<LoginResponse>, NotFound<String>> {
+#[post("/login", format = "json", data = "<data_raw>")]
+pub fn login(data_raw: Json<LoginRequest>) -> Result<Json<LoginResponse>, Status> {
     let controller: UsersController = UsersController();
-    let user_res = controller.get_user_by_name(name.into_inner().name.clone());
+    let data = data_raw.into_inner();
+    let user_check = controller.check_password(&data.name, &data.password);
+    if !user_check {
+        return Err(Status::Unauthorized);
+    }
+    let user_res = controller.get_user_by_name(&data.name);
     return match user_res {
         Ok(user) => Ok(Json(LoginResponse {
-            token: auth::create_jwt(user.user_id, user.name),
+            token: auth::create_jwt(user.user_id, &user.name),
+            user: user
         })),
         Err(err) => {
             if err.eq(&Error::NotFound) {
-                Err(NotFound(err.to_string()))
-            } else {
-                panic!("{}", err.to_string())
+                return Err(Status::Unauthorized);
             }
+            panic!("{}", err.to_string())
         }
     };
 }
@@ -84,23 +89,26 @@ fn private_endpoint(user: UserFromToken) -> String {
     return format!("{}, {}", user.0.user_id.to_string(), user.0.name);
 }
 
-// #[post("/registration", format="json", data="<name>")]
-// pub fn login(data: Json<Login>) -> Result<LoginResponse, String> {
-//     let controller : UsersController = UsersController();
-//     return Json(controller.get_user(data.name));
-// }
-
 #[get("/users")]
 fn get_users() -> Json<Vec<User>> {
     let controller: UsersController = UsersController();
     return Json(controller.get_users());
 }
 
+#[post("/users", format = "json", data = "<data_raw>")]
+fn post_users(data_raw: Json<NewUser>) -> Result<Json<User>, Status> {
+    let controller = UsersController();
+    let data = data_raw.into_inner();
 
-#[post("/users", format = "json", data = "<data>")]
-fn post_users(data: Json<NewUser>) -> Json<User> {
-    let controller: UsersController = UsersController();
-    return Json(controller.create_user(data.into_inner()));
+    match controller.get_user_by_name(&data.name) {
+        Ok(_t) => return Err(Status::Conflict),
+        Err(e) => {
+            if e.eq(&Error::NotFound) {
+                return Ok(Json(controller.create_user(data)));
+            }
+            panic!("{}", e.to_string())
+        }
+    }
 }
 
 #[get("/users/<user_id>")]
@@ -118,15 +126,17 @@ fn get_user(user_id: i32) -> Result<Json<User>, NotFound<String>> {
     };
 }
 
-
 #[put("/users/<user_id>", format = "json", data = "<data>")]
-fn put_user(user_id: i32, data: Json<UpdatedUser>) -> Result<Json<User>, NotFound<String>> {
+fn put_user(user_id: i32, data: Json<UpdatedUser>, user: UserFromToken) -> Result<Json<User>, Status> {
     let controller: UsersController = UsersController();
+    if user.0.user_id != user_id {
+        return Err(Status::Unauthorized);
+    }
     return match controller.update_user(user_id, data.into_inner()) {
         Ok(user) => Ok(Json(user)),
         Err(err) => {
             if err.eq(&Error::NotFound) {
-                Err(NotFound(err.to_string()))
+                Err(Status::NotFound)
             } else {
                 panic!("{}", err.to_string())
             }
@@ -135,17 +145,20 @@ fn put_user(user_id: i32, data: Json<UpdatedUser>) -> Result<Json<User>, NotFoun
 }
 
 #[delete("/users/<user_id>")]
-fn delete_user(user_id: i32) -> Result<Status, NotFound<String>> {
+fn delete_user(user_id: i32, user: UserFromToken) -> Result<Status, Status> {
+    if user.0.user_id != user_id {
+        return Err(Status::Unauthorized);
+    }
     let controller: UsersController = UsersController();
     return match controller.delete_user(user_id) {
         Ok(_res) => {
             if _res == 0 {
-                Err(NotFound("Not found".to_string()))
+                Err(Status::NotFound)
             } else {
                 Ok(Status::Ok)
             }
-        }
-        Err(err) => panic!("{}", err.to_string()),
+        },
+        Err(err) => panic!("{}", err.to_string())
     };
 }
 
@@ -156,26 +169,49 @@ fn get_memberships(group_id: Option<i32>, user_id: Option<i32>) -> Json<Vec<Memb
 }
 
 #[put("/memberships?<group_id>&<user_id>", format = "json", data = "<data>")]
-fn put_membership(group_id: i32, user_id: i32, data: Json<UpdatedMembership>) -> Result<Json<Membership>, NotFound<String>> {
+fn put_membership(group_id: i32, user_id: i32, data: Json<UpdatedMembership>, user: UserFromToken)
+        -> Result<Json<Membership>, Status> {
     let controller : MembershipsController = MembershipsController();
+    let my_membership = controller.get_membership(group_id, user.0.user_id);
+    match my_membership {
+        Err(e) => if e.eq(&Error::NotFound)
+        { return Err(Status::Unauthorized) } else { panic!("{}", e.to_string()) },
+        Ok(m) => if m.role != "admin" { return Err(Status::Unauthorized); }
+    }
     return match controller.update_membership(group_id, user_id, data.into_inner()) {
         Ok(membership) => Ok(Json(membership)),
         Err(err) => if err.eq(&Error::NotFound)
-        { Err(NotFound(err.to_string())) } else { panic!("{}", err.to_string()) }
+        { Err(Status::NotFound) } else { panic!("{}", err.to_string()) }
     }
 }
 
-#[post("/memberships", format = "json", data = "<data>")]
-fn post_membership(data: Json<NewMembership>) -> Json<Membership> {
+#[post("/memberships", format = "json", data = "<data_raw>")]
+fn post_membership(data_raw: Json<NewMembership>, user: UserFromToken) -> Result<Json<Membership>, Status> {
+    let data = data_raw.into_inner();
+    if user.0.user_id != data.user_id {
+        return Err(Status::Unauthorized);
+    }
     let controller : MembershipsController = MembershipsController();
-    return Json(controller.create_membership(data.into_inner()));
+    return Ok(Json(controller.create_membership(data)));
 }
 
 #[delete("/memberships?<group_id>&<user_id>")]
-fn delete_membership(group_id: i32, user_id: i32) -> Result<Status, NotFound<String>> {
+fn delete_membership(group_id: i32, user_id: i32, user: UserFromToken) -> Result<Status, Status> {
     let controller: MembershipsController = MembershipsController();
+    let my_membership = controller.get_membership(group_id, user.0.user_id);
+    match my_membership {
+        Ok(m) => {
+            if user.0.user_id != user_id && m.role != "admin" {
+                return Err(Status::Unauthorized);
+            }
+        },
+        Err(e) => return Err(Status::Unauthorized)
+    }
+    if user.0.user_id != user_id {
+        return Err(Status::Unauthorized);
+    }
     return match controller.delete_membership(group_id, user_id) {
-        Ok(_res) => if _res == 0 { Err(NotFound("Not found".to_string())) } else { Ok(Status::Ok) },
+        Ok(_res) => if _res == 0 { Err(Status::NotFound) } else { Ok(Status::Ok) },
         Err(err) => panic!("{}", err.to_string())
     };
 }
